Greatly improve the dumb algorithm to find cycles.
We now make a first depth-first search during which we find all the
cycles (but not in all directions). Then, we start a depth first
search at every vertex that is already in a cycle in order to find
all the possible directions of cycles in the graph.

Runtime decreased by about 740% for the simple_ABBA_noise test.